Move BidirectionalCollectionExtension to its own file (#582)
